+++ clients/rust/src/instructions_bincode.rs
