+++ program/benches/setup.rs
@@ -2,7 +2,7 @@ use {
     mollusk_svm_bencher::Bench,
     serde::Serialize,
     solana_config_program::state::ConfigKeys,
-    solana_config_program_client::instruction_helpers::{store, ConfigState},
+    solana_config_program_client::instructions_bincode::{store, ConfigState},
     solana_sdk::{
         account::AccountSharedData,
         hash::Hash,
